changestructure: FileHistoryGraphTest: pulled out inner classes to make them reusable
